Introduce downloads.h/cpp
